Server Statistics: move server statistics file to its own module
