+++ src/glados/api/dynamic-downloads/downloads_manager.py
@@ -8,7 +8,7 @@ from elasticsearch.helpers import scan
 from elasticsearch_dsl.connections import connections
 import json
 import traceback
-from glados import glados_server_statistics
+from glados.usage_statistics import glados_server_statistics
 from glados.api.sssearch import search_manager
 import gzip
 import os
