+++ src/glados/views.py
@@ -1,20 +1,22 @@
-from django.shortcuts import render
-from twitter import *
+import datetime
+import re
+
+import requests
+import timeago
+from apiclient.discovery import build
 from django.conf import settings
-from glados.utils import *
 from django.core.cache import cache
 from django.http import JsonResponse, HttpResponse
-from apiclient.discovery import build
-import re
+from django.shortcuts import render
 from elasticsearch_dsl import Search
-import requests
-import datetime
-import timeago
+from glados.api.url_shortening import url_shortener
+from glados.utils import *
+from twitter import *
+
+from src.glados.usage_statistics import glados_server_statistics
+from . import heatmap_helper
 from . import og_tags_generator
 from . import schema_tags_generator
-from . import glados_server_statistics
-from . import heatmap_helper
-from glados.api.url_shortening import url_shortener
 
 
 def visualise(request):
